Don't install Distribute or Pip in their own "siloed" cellars.
When installed in their own cellar, it's the equivalent of creating two more separate installations of Python. They have no clue about the Python they're supposed to be installed under, which was revealed by an error showing that PYTHONPATH was blank (when attempting to install Distribute).

Installing the packages into `/opt/boxen/homebrew/lib/python2.7/site-packages` allows the two to work as intended.