Add support for post(de)activate hooks and upgrade
This adds support for postactivate and postdeactivate in
virtualenvwrapper, as well as adding a flag to have pip do an upgrade
(in case you're not pinning your versions in your requirements files).
Also adds a pip_develop module which is meant to be used in conjunction
with a repository & virtualenv. Equivelent to "pip install -e ." or "python
setup.py develop"